Exclude jtag utils on mac

+++ dbgutils/setup.py
@@ -7,14 +7,25 @@ from os.path import expanduser
 import shutil
 
 from setuptools.command.install import install
+from sys import platform as _platform
 
 class JtagExtCommands(install):
     def run(self):
-        print('Copying jtag Codescape command scripts ....')
-        destination_path = os.path.join(expanduser("~"), "imgtec/console_scripts")
-        shutil.rmtree(destination_path, ignore_errors=True)
-        dir_util.copy_tree("jtagutils", destination_path, update=1, preserve_mode=0)
-        install.run(self)
+        print _platform
+        if _platform == "linux" or _platform == "linux2":
+            print('Copying jtag Codescape command scripts(platform:{0}'.format(_platform))
+            destination_path = os.path.join(expanduser("~"), "imgtec/console_scripts")
+            shutil.rmtree(destination_path, ignore_errors=True)
+            dir_util.copy_tree("jtagutils", destination_path, update=1, preserve_mode=0)
+            install.run(self)
+        else:
+            print('Skipping jtagutils on platform: {0}'.format(_platform))
+
+
+if _platform == "linux" or _platform == "linux2":
+    packages_list = ['probeutils', 'csrutils', 'dbgmacros', 'jtagutils']
+else:
+    packages_list = ['probeutils', 'csrutils', 'dbgmacros']
 
 setup(
     name = 'dbgutils',
@@ -22,10 +33,10 @@ setup(
     author = 'Nag Ponugoti(nag.ponugoti@fungible.com)',
     description = 'python debug utilities for F1',
     scripts = ['dbgsh'],
-    packages = find_packages(),
+    packages = packages_list,
     package_data = {'probeutils': ["probeutils/dut.cfg"]},
     include_package_data=True,
-    cmdclass={'install': JtagExtCommands},
+    cmdclass={'jtaginstall': JtagExtCommands},
     install_requires = [
         'aardvark_py',
         'argparse',
